+++ cpp/src/qsp_sim_main.cpp
@@ -74,7 +74,7 @@
 #include "qsp_sim_core/model_hooks.h"
 #include "qsp_sim_core/trajectory_writer.h"
 
-using namespace CancerVCT;
+using namespace qsp_sim_core;
 
 namespace {